Graham's files on 3 March at 17:35
This is a working program that demonstrates drawing line, empty oval, filled oval, empty rectangle and filled rectangle.
Each shape class extends JPanel.
The bounded shape classes (oval and rectangle) have two constructors. If a Color object is passed to the constructor then the bounded object is filled. If no parameters are passed to constructor then an empty shape is drawn. 
The shapes are updated in real time in the panel. When a user clicks the mouse button then it starts to draw the shape. When the user releases the mouse button then the second point of the shape is captured and the shape is drawn. The shape updates in real time as the user moves the mouse.
Currently each shape is associated with a different JPanel that is added to a JFrame. This is for demonstration purposes only.